Add files via upload
The main goal of a movie recommendation system is to predict what movies a user might like based on their previous interactions or the behavior of similar users. Machine learning algorithms analyze large amounts of user data to find patterns and similarities, enabling the system to suggest movies that match the individual user’s tastes.